Rename and refactor context package files
