+++ packages/context/README.md
@@ -13,16 +13,16 @@
 
 ```jsx
 import React from 'react';
-import { connect, ContextSystemProvider } from '@wp-g2/context';
+import { contextConnect, ContextSystemProvider } from '@wp-g2/context';
 
 // Create a React component as normal.
 const Snowman = ({ isMelting: true }) => (
 	<div>{isMelting ? 'Sad' : 'Happy'} Snowman</div>
 );
 
-// Connect your component to the Context system with the connect() higher-order function.
+// Connect your component to the Context system with the contextConnect() higher-order function.
 // Be sure to give it a memorable namespace (as the 2nd argument).
-const Olaf = connect(Snowman, 'Olaf');
+const Olaf = contextConnect(Snowman, 'Olaf');
 
 const App = () => {
 	// We can now target our <Olaf /> component by matching the Object key
