--- packages/context/src/ComponentsProvider.js
@@ -1,1 +1,0 @@
-export { ContextSystemProvider as ComponentsProvider } from './ContextSystemProvider';
