+++ packages/context/src/constants.js
@@ -1,3 +1,9 @@
 export const REACT_TYPEOF_KEY = '$$typeof';
 export const CONNECTED_NAMESPACE = 'data-g2-c16t';
 export const CONTEXT_COMPONENT_NAMESPACE = 'data-g2-c5tc8t';
+
+/**
+ * Special key where the connected namespaces are stored.
+ * This is attached to Context connected components as a static property.
+ */
+export const CONNECT_STATIC_NAMESPACE = '__contextSystemKey__';
